feat(climate) API event stream pour alimenter le front

+++ app/fr/montuelle/arduino/sensors/services/LoggerService.scala
@@ -5,7 +5,7 @@ import javax.inject.Inject
 import akka.util.ByteString
 import org.joda.time.DateTime
 import play.api.Logger
-import play.api.libs.json.Json
+import play.api.libs.json.{JsNumber, Json}
 import play.modules.reactivemongo.ReactiveMongoApi
 import play.api.libs.concurrent.Execution.Implicits.defaultContext
 import reactivemongo.play.json.collection.JSONCollection
@@ -20,7 +20,7 @@ class LoggerService @Inject() (val reactiveMongoApi: ReactiveMongoApi) {
     if (txtValue.contains("|")) {
       val values: Seq[String] = txtValue.split('|').filterNot(_.length == 0)
       Logger.debug(s"data received in text : ${values}")
-      collection.insert(Json.obj("temperature" -> values(0), "humidity" -> values(1), "time" -> new DateTime))
+      collection.insert(Json.obj("temperature" -> values(0).toInt, "humidity" -> values(1).toInt, "time" -> new DateTime))
     }
   }
